Viewer/server/requirements.txt
snowflake-connector-python[pandas]

fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
